requirements.txt
ecdsa==0.13
Flask==0.10.1
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycrypto==2.6.1
python-jose==0.6.1
six==1.10.0
Werkzeug==0.11.4
wheel==0.29.0
